requirements.txt
streamlit==1.25.0
anthropic==0.2.0
sentence-transformers==2.2.2
pinecone-client==2.2.1
transformers==4.31.0
torch==2.0.1
scikit-learn==1.2.2
numpy==1.24.2
requests==2.28.1
